Update micro_c.wdl
WDL now handles output file retrieval and cleanup (as opposed to launcher). Input re-structures to be array[struct] where the struct is now [string, array, array]. Essentially, you can have a struct for each replicate, and the input will be an array of all replicates. Within the struct, you may have multiple fastqs for an individual replicate if the library was sequenced on multiple lanes or flow cells.